+++ saleor/graphql/product/tests/test_collection.py
@@ -1,3 +1,4 @@
+import os
 from unittest.mock import Mock, patch
 
 import graphene
@@ -291,8 +292,12 @@ query CollectionProducts($id: ID!,$channel: String, $filters: ProductFilterInput
           id
           attributes {
             attribute {
-              values {
-                slug
+              choices(first: 10) {
+                edges {
+                  node {
+                    slug
+                  }
+                }
               }
             }
           }
@@ -371,7 +376,16 @@ def test_filter_collection_products_by_multiple_attributes(
         "Product", product_with_multiple_values_attributes.pk
     )
     assert product["attributes"] == [
-        {"attribute": {"values": [{"slug": "eco"}, {"slug": "power"}]}}
+        {
+            "attribute": {
+                "choices": {
+                    "edges": [
+                        {"node": {"slug": "eco"}},
+                        {"node": {"slug": "power"}},
+                    ]
+                }
+            }
+        }
     ]
 
 
@@ -449,6 +463,11 @@ def test_create_collection(
     assert data["products"]["totalCount"] == len(product_ids)
     collection = Collection.objects.get(slug=slug)
     assert collection.background_image.file
+    img_name, format = os.path.splitext(image_file._name)
+    file_name = collection.background_image.name
+    assert file_name != image_file._name
+    assert file_name.startswith(f"collection-backgrounds/{img_name}")
+    assert file_name.endswith(format)
     mock_create_thumbnails.assert_called_once_with(collection.pk)
     assert data["backgroundImage"]["alt"] == image_alt
 
@@ -685,7 +704,7 @@ def test_update_collection_invalid_background_image(
     content = get_graphql_content(response)
     data = content["data"]["collectionUpdate"]
     assert data["errors"][0]["field"] == "backgroundImage"
-    assert data["errors"][0]["message"] == "Invalid file type"
+    assert data["errors"][0]["message"] == "Invalid file type."
 
 
 UPDATE_COLLECTION_SLUG_MUTATION = """