+++ saleor/graphql/product/tests/test_product.py
@@ -1,4 +1,5 @@
 import json
+import os
 from datetime import datetime, timedelta
 from decimal import Decimal
 from unittest.mock import ANY, Mock, patch
@@ -6534,15 +6535,24 @@ PRODUCT_TYPE_CREATE_MUTATION = """
                 hasVariants
                 variantAttributes {
                     name
-                    values {
-                        name
+                    choices(first: 10) {
+                        edges {
+                            node {
+                                name
+                            }
+                        }
                     }
                 }
                 productAttributes {
                     name
-                    values {
-                        name
-                        richText
+                    choices(first: 10) {
+                        edges {
+                            node {
+                                name
+                                richText
+                            }
+                        }
+
                     }
                 }
             }
@@ -6604,15 +6614,15 @@ def test_product_type_create_mutation(
 
     pa = product_attributes[0]
     assert data["productAttributes"][0]["name"] == pa.name
-    pa_values = data["productAttributes"][0]["values"]
-    assert sorted([value["name"] for value in pa_values]) == sorted(
+    pa_values = data["productAttributes"][0]["choices"]["edges"]
+    assert sorted([value["node"]["name"] for value in pa_values]) == sorted(
         [value.name for value in pa.values.all()]
     )
 
     va = variant_attributes[0]
     assert data["variantAttributes"][0]["name"] == va.name
-    va_values = data["variantAttributes"][0]["values"]
-    assert sorted([value["name"] for value in va_values]) == sorted(
+    va_values = data["variantAttributes"][0]["choices"]["edges"]
+    assert sorted([value["node"]["name"] for value in va_values]) == sorted(
         [value.name for value in va.values.all()]
     )
 
@@ -6656,14 +6666,16 @@ def test_create_product_type_with_rich_text_attribute(
     expected_attributes = [
         {
             "name": "Color",
-            "values": [
-                {"name": "Red", "richText": None},
-                {"name": "Blue", "richText": None},
-            ],
+            "choices": {
+                "edges": [
+                    {"node": {"name": "Red", "richText": None}},
+                    {"node": {"name": "Blue", "richText": None}},
+                ]
+            },
         },
         {
             "name": "Text",
-            "values": [],
+            "choices": {"edges": []},
         },
     ]
     for attribute in data["productAttributes"]:
@@ -7334,6 +7346,11 @@ def test_product_media_create_mutation(
     product.refresh_from_db()
     product_image = product.media.last()
     assert product_image.image.file
+    img_name, format = os.path.splitext(image_file._name)
+    file_name = product_image.image.name
+    assert file_name != image_file._name
+    assert file_name.startswith(f"products/{img_name}")
+    assert file_name.endswith(format)
 
     # The image creation should have triggered a warm-up
     mock_create_thumbnails.assert_called_once_with(product_image.pk)
@@ -7493,7 +7510,7 @@ def test_invalid_product_media_create_mutation(
 
     content = get_graphql_content(response)
     assert content["data"]["productMediaCreate"]["errors"] == [
-        {"field": "image", "message": "Invalid file type"}
+        {"field": "image", "message": "Invalid file type."}
     ]
     product.refresh_from_db()
     assert product.media.count() == 0